:hammer: refac: contexts to ts
