+++ package.json
@@ -6,6 +6,7 @@
   "author": "Henrique Miranda <enrque.ns@gmail.com>",
   "dependencies": {
     "@apollo/react-hooks": "^3.1.3",
+    "@types/isomorphic-fetch": "^0.0.35",
     "apollo-boost": "^0.4.7",
     "babel-plugin-polished": "^1.1.0",
     "babel-plugin-styled-components": "^1.10.6",
