--- src/context/SectionsContext.js
@@ -1,34 +1,0 @@
-import React, { createContext, useState } from 'react';
-
-import { useViewportScroll } from 'framer-motion';
-import PropTypes from 'prop-types';
-
-const SectionsContext = createContext();
-export const sections = ['home', 'education', 'skills', 'projects', 'about'];
-
-export function SectionsProvider({ children }) {
-  const [selected, setSelectedByName] = useState('home');
-  const [pageHeight, setPageHeight] = useState(0);
-  const { scrollY } = useViewportScroll();
-
-  return (
-    <SectionsContext.Provider
-      value={{
-        scrollY,
-        selected,
-        sections,
-        setSelectedByName,
-        setPageHeight,
-        pageHeight,
-      }}
-    >
-      {children}
-    </SectionsContext.Provider>
-  );
-}
-
-SectionsProvider.propTypes = {
-  children: PropTypes.node.isRequired,
-};
-
-export default SectionsContext;
